telegram-bot/requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-dateutil==2.8.2
pytest==7.4.4
pytest-asyncio==0.23.3
